requirements.txt
# Python package dependencies for AI Proxy News
google-cloud-speech>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
google-cloud-storage>=2.0.0
websockets>=11.0
